requirements.txt
requests==2.22.0
Naked==0.1.31
pychalk==2.0.1